requirements.txt
# --- 核心框架 (FastAPI & Streamlit) ---
fastapi==0.128.0
uvicorn[standard]==0.40.0
streamlit==1.53.1
extra-streamlit-components==0.1.81
streamlit-searchbox==0.1.24

# --- 数据库 & ORM (人大金仓兼容相关) ---
SQLAlchemy==2.0.46
asyncpg==0.31.0
psycopg2-binary==2.9.11
psycopg[binary,pool]==3.3.2
pg8000==1.31.5
greenlet==3.3.1

# --- 数据处理 & 科学计算 ---
pandas==2.3.3
numpy==2.4.2
scipy==1.17.1
scikit-learn==1.8.0
pyarrow==23.0.0

# --- 身份验证 & 工具 ---
fastapi-users[sqlalchemy]==15.0.3
python-dotenv==1.2.1
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings==2.12.0
PyJWT==2.10.1
bcrypt==3.2.2
passlib[bcrypt]==1.7.4

# --- 网络请求 & API ---
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3
imagekitio==5.2.0

# --- 其它工具 ---
Pygments==2.19.2
rich==14.3.2
watchdog==6.0.0
jinja2==3.1.6
pyyaml==6.0.3
GitPython==3.1.46